pyproject.toml
[project]
name = "mcc"
version = "0.1.0"
description = "Mission Control Center — multi-agent AI software development platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyjwt>=2.10.0",
    "pwdlib[argon2,bcrypt]>=0.2.1",
    "pyotp>=2.9.0",
    "python-multipart>=0.0.18",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.28.0",
    "email-validator>=2.2.0",
    "python-socketio>=5.11.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
